src/requirements.txt
fastapi
uvicorn
openai
pydantic
scikit-learn
Pillow
python-dotenv
